requirements.txt
tweepy
apscheduler
supabase
python-dotenv
